requirements.txt
adal==1.2.7
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-tools==0.2
applicationinsights==0.11.10
appnope==0.1.3
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
azure-cli-core==2.11.1
azure-cli-nspkg==3.0.2
azure-cli-telemetry==1.0.8
azure-common==1.1.11
azure-core==1.26.1
azure-graphrbac==0.61.1
azure-keyvault==1.0.0a1
azure-mgmt-authorization==0.51.1
azure-mgmt-automation==0.1.1
azure-mgmt-batch==5.0.1
azure-mgmt-cdn==3.0.0
azure-mgmt-compute==10.0.0
azure-mgmt-containerinstance==1.4.0
azure-mgmt-containerregistry==2.0.0
azure-mgmt-containerservice==9.1.0
azure-mgmt-core==1.2.0
azure-mgmt-cosmosdb==0.5.2
azure-mgmt-devtestlabs==3.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-hdinsight==0.1.0
azure-mgmt-iothub==0.7.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==1.0.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-monitor==0.5.2
azure-mgmt-network==10.2.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-redis==5.0.0
azure-mgmt-resource==10.2.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-sql==0.10.0
azure-mgmt-storage==11.1.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.41.0
azure-nspkg==2.0.0
azure-storage==0.35.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.3
cycler==0.11.0
datasets==2.10.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
entrypoints==0.4
environs==9.5.0
evaluate==0.4.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
fonttools==4.37.4
frozenlist==1.3.3
fsspec==2023.1.0
gdown==4.6.4
googleapis-common-protos==1.58.0
grpcio==1.51.1
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpie==3.2.1
huggingface-hub==0.12.1
humanfriendly==8.2
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
install==1.3.5
install-jdk==1.0.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
knack==0.7.2
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
msal==1.0.0
msal-extensions==0.1.3
msrest==0.6.10
msrestazure==0.6.4
multidict==6.0.2
multiprocess==0.70.14
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nlpaug==1.1.11
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.2
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-instrumentation==0.37b0
opentelemetry-instrumentation-aws-lambda==0.37b0
opentelemetry-instrumentation-dbapi==0.37b0
opentelemetry-instrumentation-grpc==0.37b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
opentelemetry-util-http==0.37b0
optparse-pretty==0.1.1
packaging==23.0
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
perlin-noise==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==3.0.0
portalocker==1.7.1
priority==2.0.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pygame==2.1.2
pyglet==2.0.3
Pygments==2.14.0
PyJWT==1.7.1
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.3.2
QtPy==2.2.1
quart==0.18.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rich==12.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
terminado==0.16.0
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.1.1
tokenizers==0.13.2
toml==0.10.2
torch==2.0.0.dev20230226
torchaudio==2.0.0.dev20230226
torchvision==0.15.0.dev20230226
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.8.2
